Use an example HTML closer to the voter app
It was suggested that showcasing HTML with a structure and content which
is closer to the one that will be produced later in the voting app will
be more beneficial than showcasing the simple profile page.

I still think that the profile page might have appeal to some
participants since we humans are naturally narcissistic and since the
page is simpler, so I kept it as an alternative.